src/test/javascript/spec/app/entities/quan-he-than-nhan/quan-he-than-nhan.service.spec.ts
/* tslint:disable max-line-length */
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { HttpClient, HttpResponse } from '@angular/common/http';
import { of } from 'rxjs';
import { take, map } from 'rxjs/operators';
import { QuanHeThanNhanService } from 'app/entities/quan-he-than-nhan/quan-he-than-nhan.service';
import { IQuanHeThanNhan, QuanHeThanNhan } from 'app/shared/model/quan-he-than-nhan.model';

describe('Service Tests', () => {
    describe('QuanHeThanNhan Service', () => {
        let injector: TestBed;
        let service: QuanHeThanNhanService;
        let httpMock: HttpTestingController;
        let elemDefault: IQuanHeThanNhan;
        beforeEach(() => {
            TestBed.configureTestingModule({
                imports: [HttpClientTestingModule]
            });
            injector = getTestBed();
            service = injector.get(QuanHeThanNhanService);
            httpMock = injector.get(HttpTestingController);

            elemDefault = new QuanHeThanNhan(0, 'AAAAAAA', 'AAAAAAA', 'AAAAAAA', false);
        });

        describe('Service methods', async () => {
            it('should find an element', async () => {
                const returnedFromService = Object.assign({}, elemDefault);
                service
                    .find(123)
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: elemDefault }));

                const req = httpMock.expectOne({ method: 'GET' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should create a QuanHeThanNhan', async () => {
                const returnedFromService = Object.assign(
                    {
                        id: 0
                    },
                    elemDefault
                );
                const expected = Object.assign({}, returnedFromService);
                service
                    .create(new QuanHeThanNhan(null))
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: expected }));
                const req = httpMock.expectOne({ method: 'POST' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should update a QuanHeThanNhan', async () => {
                const returnedFromService = Object.assign(
                    {
                        maQuanHe: 'BBBBBB',
                        loaiQuanHe: 'BBBBBB',
                        moTa: 'BBBBBB',
                        isDeleted: true
                    },
                    elemDefault
                );

                const expected = Object.assign({}, returnedFromService);
                service
                    .update(expected)
                    .pipe(take(1))
                    .subscribe(resp => expect(resp).toMatchObject({ body: expected }));
                const req = httpMock.expectOne({ method: 'PUT' });
                req.flush(JSON.stringify(returnedFromService));
            });

            it('should return a list of QuanHeThanNhan', async () => {
                const returnedFromService = Object.assign(
                    {
                        maQuanHe: 'BBBBBB',
                        loaiQuanHe: 'BBBBBB',
                        moTa: 'BBBBBB',
                        isDeleted: true
                    },
                    elemDefault
                );
                const expected = Object.assign({}, returnedFromService);
                service
                    .query(expected)
                    .pipe(
                        take(1),
                        map(resp => resp.body)
                    )
                    .subscribe(body => expect(body).toContainEqual(expected));
                const req = httpMock.expectOne({ method: 'GET' });
                req.flush(JSON.stringify([returnedFromService]));
                httpMock.verify();
            });

            it('should delete a QuanHeThanNhan', async () => {
                const rxPromise = service.delete(123).subscribe(resp => expect(resp.ok));

                const req = httpMock.expectOne({ method: 'DELETE' });
                req.flush({ status: 200 });
            });
        });

        afterEach(() => {
            httpMock.verify();
        });
    });
});
